Added skill-based agility calculations. Agility now uses getters and setters.

+++ src/disparity/skills/Fight.java
@@ -7,7 +7,7 @@ public class Fight{
 	public static byte doesHit(Being a, Weapon w, Being b){
 		int i = 0;
 		double max = 100.0;
-		double min = w.base_dmg + (int)(w.getSWDB(a).damBonus);
+		double min = w.base_dmg + (w.getSWDB(a).damBonus);
 		if(min + Math.random() * ((max - min) + 1) <= b.calcAC())	i += 1; 		//checks to see if target blocked the attack [1 = block]
 		else if(min + Math.random() * ((max - min) + 1) <= b.dodge)	i += 2; //checks to see if target dodged the attack [2 = dodge]
 																			//if target does not block or dodge, attack hits [default = hit]
